train | refactor
* move prepare_run to own module
